fix: Stop click propagation for `Modal` component (#714)
# Motivation

When migrating component Modal to Svelte 5, there was an oversight in
the propagation of the `on:click` event.

### Before migration

```ts
  <div
    ...
    on:click|stopPropagation
    ...
  >
```

### After migration (incorrect)

```ts
  <div
    ...
    onclick={nonNullish(onClick) ? stopPropagation(onClick) : undefined}
    ...
  >
```


This version is not correct, since it does not block the propagation if
`onClick` is nullish. So it is correct this other version:

### After migration (correct)

```ts
  <div
    ...
    onclick={stopPropagation(nonNullish(onClick) ?  onClick : ()=> {}) }
    ...
  >
```

# Tests

Created a new test. Such a test would be failing before this PR, as
tested locally.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>