chore: updates workflow to enable npm trusted publishing (#724)
# Motivation

NPM package publishing will be migrated to use Trusted Publishing (via
OIDC).

To ensure security and restrict access, we will configure the OIDC token
request to only be available to jobs running in the dedicated 'release'
environment. This environment will be a required condition for obtaining
the necessary NPM token.

The consolidation into one file is required, because npmjs.org does not
support multiple filenames that can obtain a token.

# Changes

* Consolidate workflows `publish.yml` and `next.yml` into one file.
* Bump npm version to v11 to support OIDC.
* Add `release` environment to be able to obtain the short-lived token.
* Removed the static NPM token.