chore: enforce npm >=11.5.1 (#727)
# Motivation

npm is deprecting NPM_TOKEN for publishing and is introducing the
support for trusted publisher. We are migrating to this approach in the
CI which requires npm >= v11.5.1. That is why, similarly e.g. as in the
OISY Wallet Signer, I propose to set this version as mandatory for
development as well. This way we avoid conflicts in `package-lock.json`.

# Notes

Follow-up of https://github.com/dfinity/gix-components/pull/724

# Changes

- Set a field `engines` in `package.json` to define npm and node version
(LTS here)