chore: strip engines field for release (#726)
# Motivation

We want to enforce our development environment and CI to use npm v11
but, we do not want consumers and devs to use this version. That is why
we strip `engines` before release from the tarball we publish to npm.

# Notes

Follow-up of https://github.com/dfinity/gix-components/pull/724 and #727

# Changes

- Add a script to strip and redo the `engines` field. It uses a backup
file to redo the `package.json` after pack in case we would ever do this
locally. The library is published with its root `package.json`. Note
that I did not added a git exception as the backup file is deleted if
everything works well.